+++ src/Scripts/FunctionMappings/HeavyRanged.lua
@@ -1,68 +1,67 @@
--- Functions required by the HeavyRanged enemy
-
 -- Creates "Tethers", which are floating parts of the enemy (e.g. the small crystals for HeavyRanged or the neck for the Hydra heads)
 function mod.CreateTethers(newEnemy, args)
-	if newEnemy == nil or newEnemy.Tethers == nil or newEnemy.TetherIds ~= nil then
-		return
-	end
+	-- Tethers have been causing almost all of the currently reported crashes
+	-- The issue appears to be the small tethered Crystals causing touchdowner is nil errors in OnTouchdown
+	-- So for now, we are disabling tethers entirely
+	-- This is tracked in #329 
+	return
 
-	-- For current compatibility with Gale, if the player has her equipped, don't create Tethers as they might be what's causing the crashes when she attacks them #329
-	if game.GameState.EquippedFamiliar == "PolecatFamiliar" or game.GameState.EquippedFamilar == "FrogFamiliar" then
-		return
-	end
+	-- if newEnemy == nil or newEnemy.Tethers == nil or newEnemy.TetherIds ~= nil then
+	-- 	return
+	-- end
 
-	newEnemy.TetherIds = {}
-	local prevTetherId = newEnemy.ObjectId
-	for k, tether in ipairs(newEnemy.Tethers) do
-		local count = tether.Count or 1
-		for i = 1, count do
-			local offsetX = 0.0
-			local offsetY = 0.0
-			if tether.SpawnRadius ~= nil then
-				offsetX = game.RandomFloat(-tether.SpawnRadius, tether.SpawnRadius) or 0.0
-				offsetY = game.RandomFloat(-tether.SpawnRadius, tether.SpawnRadius) or 0.0
-			end
-			local tetherId = SpawnObstacle({
-				Name = tether.Name,
-				DestinationId = newEnemy.ObjectId,
-				Group = tether.GroupName or "Standing",
-				OffsetX = offsetX,
-				OffsetY = offsetY
-			})
-			SetAlpha({ Id = tetherId, Fraction = 0 })
-			SetAlpha({ Id = tetherId, Fraction = 1.0, Duration = 0.3 })
-			if tether.Elasticity ~= nil then
-				Attach({
-					Id = tetherId,
-					DestinationId = newEnemy.ObjectId,
-					-- The issue with the tethers is that TetherDistance and TetherElasticity are no longer existent in Hades II - issue #22
-					TetherDistance = tether.Distance,
-					TetherElasticity = tether.Elasticity,
-					-- As a workaround, just do some static offsets
-					OffsetX = offsetX,
-					OffsetY = offsetY,
-				})
-			else
-				Attach({
-					Id = prevTetherId,
-					DestinationId = tetherId,
-					TetherDistance = tether.Distance,
-					TetherRetractSpeed = tether.RetractSpeed,
-					TetherTrackZRatio = tether.TrackZRatio,
-					OffsetX = offsetX,
-					OffsetY = offsetY,
-				})
-			end
-			table.insert(newEnemy.TetherIds, tetherId)
-			if (newEnemy.EliteIcon or (newEnemy.HealthBuffer ~= nil and newEnemy.HealthBuffer > 0)) and tetherId ~= newEnemy.ObjectId then
-				newEnemy.Outline.Id = tetherId
-				if newEnemy.Outline.Thickness > 0 then
-					AddOutline(newEnemy.Outline)
-				end
-			end
-			prevTetherId = tetherId
-		end
-	end
+	-- newEnemy.TetherIds = {}
+	-- local prevTetherId = newEnemy.ObjectId
+	-- for k, tether in ipairs(newEnemy.Tethers) do
+	-- 	local count = tether.Count or 1
+	-- 	for i = 1, count do
+	-- 		local offsetX = 0.0
+	-- 		local offsetY = 0.0
+	-- 		if tether.SpawnRadius ~= nil then
+	-- 			offsetX = game.RandomFloat(-tether.SpawnRadius, tether.SpawnRadius) or 0.0
+	-- 			offsetY = game.RandomFloat(-tether.SpawnRadius, tether.SpawnRadius) or 0.0
+	-- 		end
+	-- 		local tetherId = SpawnObstacle({
+	-- 			Name = tether.Name,
+	-- 			DestinationId = newEnemy.ObjectId,
+	-- 			Group = tether.GroupName or "Standing",
+	-- 			OffsetX = offsetX,
+	-- 			OffsetY = offsetY
+	-- 		})
+	-- 		SetAlpha({ Id = tetherId, Fraction = 0 })
+	-- 		SetAlpha({ Id = tetherId, Fraction = 1.0, Duration = 0.3 })
+	-- 		if tether.Elasticity ~= nil then
+	-- 			Attach({
+	-- 				Id = tetherId,
+	-- 				DestinationId = newEnemy.ObjectId,
+	-- 				-- The issue with the tethers is that TetherDistance and TetherElasticity are no longer existent in Hades II - issue #22
+	-- 				TetherDistance = tether.Distance,
+	-- 				TetherElasticity = tether.Elasticity,
+	-- 				-- As a workaround, just do some static offsets
+	-- 				OffsetX = offsetX,
+	-- 				OffsetY = offsetY,
+	-- 			})
+	-- 		else
+	-- 			Attach({
+	-- 				Id = prevTetherId,
+	-- 				DestinationId = tetherId,
+	-- 				TetherDistance = tether.Distance,
+	-- 				TetherRetractSpeed = tether.RetractSpeed,
+	-- 				TetherTrackZRatio = tether.TrackZRatio,
+	-- 				OffsetX = offsetX,
+	-- 				OffsetY = offsetY,
+	-- 			})
+	-- 		end
+	-- 		table.insert(newEnemy.TetherIds, tetherId)
+	-- 		if (newEnemy.EliteIcon or (newEnemy.HealthBuffer ~= nil and newEnemy.HealthBuffer > 0)) and tetherId ~= newEnemy.ObjectId then
+	-- 			newEnemy.Outline.Id = tetherId
+	-- 			if newEnemy.Outline.Thickness > 0 then
+	-- 				AddOutline(newEnemy.Outline)
+	-- 			end
+	-- 		end
+	-- 		prevTetherId = tetherId
+	-- 	end
+	-- end
 end
 
 function mod.HandleTetherParentDeath(victim, skipTetherCount, skipTetherAnimation)